banking-platform/account-agent-a2a/a2a-agent/pyproject.toml
[project]
name = "account-agent-a2a"
version = "1.0.0"
description = "A2A-compatible Account Agent using Azure AI Foundry and MCP"
requires-python = ">=3.11"
dependencies = [
    # Agent Framework (PyPI packages)
    "agent-framework-core==1.0.0b251223",
    "agent-framework-azure-ai==1.0.0b251223",
    "agent-framework-a2a==1.0.0b251223",
    
    # Azure AI Foundry SDK
    "azure-ai-projects>=2.0.0b2",
    "azure-ai-agents==1.2.0b5",
    
    # A2A SDK
    "a2a-sdk>=0.3.5",
    
    # Azure Identity
    "azure-identity>=1,<2",
    
    # Pydantic
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    
    # OpenAI SDK
    "openai>=1.99.0",
    
    # MCP support
    "mcp[ws]>=1.23",
    
    # FastAPI for A2A HTTP server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    
    # HTTP client
    "aiohttp",
    "httpx>=0.27.0",
    
    # Environment variables
    "python-dotenv>=1.0.0",
    
    # Utilities
    "typing-extensions",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
